refactor: extract video event listeners to content/videoListeners.ts
- Move setupVideoListeners, videoOnReadyListener, durationChangeListener
  and associated variables (playbackRateCheckInterval, lastPlaybackSpeed,
  setupVideoListenersFirstTime) into new content/videoListeners module
- Use dependency injection pattern (initVideoListeners) for
  updateVisibilityOfPlayerControlsButton, consistent with other modules
- Remove now-unused imports from content.ts (danmakuForSkip,
  cancelSponsorSchedule, clearWaitingTime, updateVirtualTime,
  updateWaitingTime, addCleanupListener)
